worker/requirements.txt
flask==3.0.3
gunicorn==21.2.0
deface==1.5.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
supabase==2.4.6
python-dotenv==1.0.1
numpy==1.26.4
